% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{binary_edge_histogram}
\alias{binary_edge_histogram}
\title{plot histogram of the distribution of edges}
\usage{
binary_edge_histogram(edge_lists, og_edges, xlimits = c(0, 1),
  title = "", color = "lightskyblue")
}
\arguments{
\item{edge_lists}{list of vectors of boolean elements indicating which sets have edges between them}

\item{og_edges}{vector of boolean values indicating the observed edges}

\item{xlimits}{x-axis range}

\item{title}{title of plot}

\item{color}{color of bars}
}
\value{
NA
}
\description{
plot histogram of the distribution of edges
}
\examples{

}
