% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{binary_set_assignment}
\alias{binary_set_assignment}
\title{convert sets to groups of binary integers depending on whether or not each element fits certain criteria}
\usage{
binary_set_assignment(sets, assignment)
}
\arguments{
\item{sets}{original sets}

\item{assignment}{list of elements which fit the criteria (these will be set to 1 and the remaining will be set to 0)}
}
\value{
list of sets in which elements are either 1 or 0
}
\description{
convert sets to groups of binary integers depending on whether or not each element fits certain criteria
}
\examples{

}
