% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{binary_set_histogram}
\alias{binary_set_histogram}
\title{calculate the fraction of each binary set whose elements are 1}
\usage{
binary_set_histogram(binary_sets)
}
\arguments{
\item{binary_sets}{list of n sets whose elements are 1 or 0}
}
\value{
n by 1 matrix of doubles
}
\description{
calculate the fraction of each binary set whose elements are 1
}
\examples{

}
