% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{bootstrap_dfde}
\alias{bootstrap_dfde}
\title{shuffle df and de set fractions and observe overlap to assign dfde}
\usage{
bootstrap_dfde(df_frac_list, de_frac_list, set_size_list, fracs = TRUE)
}
\arguments{
\item{df_frac_list}{list of set-fractions of df genes}

\item{de_frac_list}{list of set-frctions of de genes}

\item{set_size_list}{list of integers that indicate the number of elements in each set}

\item{fracs}{boolean value indicating whether or not to return the fraction or number of bootstrapped dfde genes in each set. If TRUE, will return fractions.}
}
\value{
list of integers or doubles indicating the bootstrapped fraction or number of dfde genes in each set
}
\description{
shuffle df and de set fractions and observe overlap to assign dfde
}
\examples{

}
