% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{clean_sets}
\alias{clean_sets}
\title{clean names of genes in sets for easier analysis}
\usage{
clean_sets(sets, pattern = "PA\\\\d{4}")
}
\arguments{
\item{sets}{sets}

\item{pattern}{character pattern to clean}
}
\value{
sets with cleaned names
}
\description{
clean names of genes in sets for easier analysis
}
\examples{

}
