% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{draw_heatmap_g0}
\alias{draw_heatmap_g0}
\title{draw a heatmap of correlation values between genes in a specified G set and a G(R) set containing it}
\usage{
draw_heatmap_g0(g0_idx, gr0_idx, g0, gr0, C)
}
\arguments{
\item{g0_idx}{integer indicating the index of the G set of interest}

\item{gr0_idx}{integer indicating the index of the G(R) set of interest}

\item{C}{gene expression correlation matrix}
}
\value{
NA
}
\description{
draw a heatmap of correlation values between genes in a specified G set and a G(R) set containing it
}
\examples{

}
