% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{g1_architecture_bootstrap}
\alias{g1_architecture_bootstrap}
\title{boostrap the distribution of sets in the established G* architecture}
\usage{
g1_architecture_bootstrap(coupling_mtx, df_frac, de_frac, n = 1000,
  dfde_var = 5)
}
\arguments{
\item{coupling_mtx}{matrix indicating coupling between G sets}

\item{df_frac}{vector indicating the fraction of df genes in each set}

\item{de_frac}{vector indicating the fraction of de genes in each set}

\item{n}{number of bootstrap iterations to sample. Default is 1000}
}
\value{
matrix of boolean values in which each row indicates the edge which is present of absent and each column represents a bootstrapped iteration and sampling variant
}
\description{
boostrap the distribution of sets in the established G* architecture
}
\examples{

}
