% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{g1_architecture_measurement_binary}
\alias{g1_architecture_measurement_binary}
\title{count the number of edges between df, de, and dfde sets}
\usage{
g1_architecture_measurement_binary(coupling_mtx, df_frac, de_frac,
  dfde_var)
}
\arguments{
\item{coupling_mtx}{matrix of boolean values indicating edges between sets indicated by rows and columns}

\item{df_frac}{list of doubles indicating the fraction of df genes in each set}

\item{de_frac}{list of doubles indicating the fraction of de genes in each set}

\item{dfde_var}{integer indicating how dfde edges should be interpreted [1 = exactly one set is dfde, 2 = at least one set is dfde, 3 = neither may be dfde, 4 = neither set is dfde, 5 = both sets are dfde, anything else defaults to 1]}
}
\value{
n by 3 matrix (n being the total number of edges). First column indicates edges between df sets, second column indicates edges between de sets, third column indicates edges between df and de
}
\description{
count the number of edges between df, de, and dfde sets
}
\examples{

}
