% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{gene_data_frame}
\alias{gene_data_frame}
\title{create a data frame for making graphs}
\usage{
gene_data_frame(model, gr0, gr1)
}
\arguments{
\item{model}{sybil model}

\item{gr0}{g(r) sets}

\item{gr1}{g(r*) sets}
}
\value{
data frame
}
\description{
create a data frame for making graphs
}
\examples{

}
