% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{gene_set_dataframe}
\alias{gene_set_dataframe}
\title{generate a data frame for use in analysis from gene sets}
\usage{
gene_set_dataframe(sets)
}
\arguments{
\item{set}{G sets or G(R) sets}
}
\value{
data frame
}
\description{
generate a data frame for use in analysis from gene sets
}
\examples{

}
