% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{gene_set_expanded_dataframe}
\alias{gene_set_expanded_dataframe}
\title{expand a dataframe of gene sets to include binary sets indicating genes which show differential fitness, differential expression, either, purely one or the other, or both}
\usage{
gene_set_expanded_dataframe(set_df, sets, df_genes, de_genes,
  interest_genes, pure_df_genes, pure_de_genes, dfde_genes)
}
\arguments{
\item{set_df}{dataframe of gene sets}

\item{sets}{gene sets}

\item{df_genes}{list of genes which show differential fitness (df) effects}

\item{de_genes}{list of genes which show differential expression (de)}

\item{interest_genes}{list of genes which show differential fitness or expression}

\item{pure_df_genes}{list of genes which show df but not de}

\item{pure_de_genes}{list of genes which show de but not df}

\item{dfde_genes}{list of genes which show df and de effects}
}
\value{
dataframe of sets and binary sets for all listed criteria
}
\description{
expand a dataframe of gene sets to include binary sets indicating genes which show differential fitness, differential expression, either, purely one or the other, or both
}
\examples{

}
