% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{map_genes_to_sets}
\alias{map_genes_to_sets}
\title{return a matrix indicating the index of the set at which each gene can be found}
\usage{
map_genes_to_sets(genes, set)
}
\arguments{
\item{genes}{list of n genes of interest}

\item{set}{list of sets}
}
\value{
n by 1 matrix of integers
}
\description{
return a matrix indicating the index of the set at which each gene can be found
}
\examples{

}
