% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{plot_density}
\alias{plot_density}
\title{plot boostrapped values to assess the spread of set completeness}
\usage{
plot_density(mtx, xlimits = c(0, 1), ylimits = c(0, 200), start = 2)
}
\arguments{
\item{mtx}{matrix of bootstrapped set-fraction values}

\item{xlimits}{range of x-axis for plot}

\item{ylimits}{range of y-axis for plot}

\item{start}{first column to plot (will plot from start to ncol). the first column is a list of zeros so the default value is 2}
}
\value{
NA
}
\description{
plot boostrapped values to assess the spread of set completeness
}
\examples{

}
