% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{plot_density_compar}
\alias{plot_density_compar}
\title{plot boostrapped values to assess the spread of set completeness against observed values}
\usage{
plot_density_compar(mtx, xlimits = c(0, 1), ylimits = c(0, 200),
  start = 2, freqs)
}
\arguments{
\item{mtx}{matrix of bootstrapped set-fraction values}

\item{xlimits}{range of x-axis for plot}

\item{ylimits}{range of y-axis for plot}

\item{start}{first column to plot (will plot from start to ncol). the first column is a list of zeros so the default value is 2}

\item{freqs}{table of observed values}
}
\value{
NA
}
\description{
plot boostrapped values to assess the spread of set completeness against observed values
}
\examples{

}
