% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{plot_dfde_bootstrap}
\alias{plot_dfde_bootstrap}
\title{bootstrap the distribution of dfde genes in sets by only shuffling the assignments within sets and plot}
\usage{
plot_dfde_bootstrap(df)
}
\arguments{
\item{df}{dataframe of gene set information}
}
\value{
NA
}
\description{
bootstrap the distribution of dfde genes in sets by only shuffling the assignments within sets and plot
}
\examples{

}
