% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{plot_dfde_sampling}
\alias{plot_dfde_sampling}
\title{bootstrap and plot the distribution of dfde genes}
\usage{
plot_dfde_sampling(sets, num_df, num_de, num_dfde)
}
\arguments{
\item{sets}{original gene sets}

\item{num_df}{number of df genes}

\item{num_de}{number of de genes}

\item{num_dfde}{original number of dfde genes}
}
\value{
NA
}
\description{
bootstrap and plot the distribution of dfde genes
}
\examples{

}
