% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{plot_in_out}
\alias{plot_in_out}
\title{plot correlation density of sets}
\usage{
plot_in_out(sets, C = C)
}
\arguments{
\item{sets}{sets to analyze correlation of}

\item{C}{correlation matric}
}
\value{
NA
}
\description{
plot correlation density of sets
}
\examples{

}
