% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{plot_set_characteristic_hypothesis}
\alias{plot_set_characteristic_hypothesis}
\title{plot the number of sets which are fully complete or empty}
\usage{
plot_set_characteristic_hypothesis(observed, simulated, start = 2,
  end = 1001, title = "", color = "lightskyblue")
}
\arguments{
\item{observed}{list of observed set fractions}

\item{simulated}{matrix of bootstrapped sete fractions}

\item{start}{first column of 'simulated' to plot (will plot from start to ncol). the first column is a list of zeros so the default value is 2}

\item{end}{last column of 'simulated' to plot. the first column is a list of zeros so the default value is 1001 to hvae 1000 bootstrapped values}

\item{title}{plot title}

\item{color}{color for histogram bars}
}
\value{
NA
}
\description{
plot the number of sets which are fully complete or empty
}
\examples{

}
