% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{sample_dfde_sets}
\alias{sample_dfde_sets}
\title{function for one boostrap iteration for the distribution of dfde genes in gene sets}
\usage{
sample_dfde_sets(sets, n_df, n_de)
}
\arguments{
\item{sets}{list of genes sets}

\item{n_df}{number of df genes to randomly assign}

\item{n_de}{number of de genes to randomly assign}
}
\value{
list of bootstrapped dfde genes
}
\description{
function for one boostrap iteration for the distribution of dfde genes in gene sets
}
\examples{

}
