% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{set_analysis}
\alias{set_analysis}
\title{calculate average correlation values in all sets as well as quantiles}
\usage{
set_analysis(sets, C)
}
\arguments{
\item{sets}{list of sets}

\item{C}{gene expression correlation matrix}
}
\value{
dataframe of various analyses of the sets
}
\description{
calculate average correlation values in all sets as well as quantiles
}
\examples{

}
