% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{set_cor}
\alias{set_cor}
\title{calculate average correlation values between pairs in each set}
\usage{
set_cor(set, avg = TRUE, C = C)
}
\arguments{
\item{set}{sets to extract pairs from for correlations}

\item{C}{correlation matrix}
}
\value{
list of integers
}
\description{
calculate average correlation values between pairs in each set
}
\examples{

}
